+++ app/src/main/res/layout/fragment_first.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Modele.MVC.FirstFragment">
+    tools:context=".Modele.MVC.presentation.view.FirstFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler_view"
