+++ app/src/main/res/layout/log_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Modele.MVC.LogActivity">
+    tools:context=".Modele.MVC.presentation.view.LogActivity">
 
     <ProgressBar
         android:id="@+id/progressBar"
