+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/FirstFragment"
-        android:name="com.example.leguidebienpnard.Modele.MVC.FirstFragment"
+        android:name="com.example.leguidebienpnard.Modele.MVC.presentation.view.FirstFragment"
         android:label="@string/first_fragment_label"
         tools:layout="@layout/fragment_first">
 
@@ -21,7 +21,7 @@
     </fragment>
     <fragment
         android:id="@+id/SecondFragment"
-        android:name="com.example.leguidebienpnard.Modele.MVC.SecondFragment"
+        android:name="com.example.leguidebienpnard.Modele.MVC.presentation.view.SecondFragment"
         android:label="@string/second_fragment_label"
         tools:layout="@layout/fragment_second">
